backend/services/test_service.py
# Linux Mirror Testing Solution - Test Service

import asyncio
import logging
import yaml
from typing import Dict, Any
from datetime import datetime
from models.test import TestResult, TestRequest

logger = logging.getLogger(__name__)

class TestService:
    def __init__(self):
        self.running_tests = {}
        
        # Version to codename mapping for Debian and Ubuntu
        self.version_codename_mapping = {
            # Debian versions
            "7": "wheezy",
            "8": "jessie", 
            "9": "stretch",
            "10": "buster",
            "11": "bullseye",
            "12": "bookworm",
            "13": "trixie",
            # Ubuntu versions
            "18.04": "bionic",
            "20.04": "focal",
            "22.04": "jammy",
            "24.04": "noble",
            "25.04": "oracular"
        }
        
        # Load configuration
        try:
            with open('config.yaml', 'r') as file:
                self.config = yaml.safe_load(file)
        except Exception as e:
            print(f"Error reading config file: {e}")
            self.config = {}
    
    def get_codename_for_version(self, distribution: str, version: str) -> str:
        """
        Get the codename for a given distribution and version
        """
        # For Debian and Ubuntu, we need to map versions to codenames
        if distribution in ["debian", "ubuntu"]:
            return self.version_codename_mapping.get(version, version)
        else:
            # For other distributions, return the version as is
            return version
    
    async def run_test(self, test_request: TestRequest) -> TestResult:
        """
        Run a test for the specified distributions and repositories
        """
        # In a real implementation, this would:
        # 1. Spin up containers for each distro/version
        # 2. Configure repository pointing to air-gapped mirror
        # 3. Attempt package updates and installations
        # 4. Validate package integrity
        # 5. Return results
        
        logger.info(f"Starting test for distributions: {test_request.distributions}")
        
        # Simulate a test run
        await asyncio.sleep(2)
        
        # For now, we'll use generic values in the result since this is a mock implementation
        result = TestResult(
            id="test-123",
            distribution="ubuntu-20.04",  # This should be dynamic based on test request
            repository="http://mirror.example.com/ubuntu",  # This should be from config
            status="success",
            start_time=datetime.now(),
            end_time=datetime.now(),
            duration_seconds=2,
            error_message=None,
            package_count=100,
            test_details={"packages": ["curl", "wget", "git"]}
        )
        
        return result
    
    async def get_test_status(self, test_id: str) -> Dict[str, Any]:
        """
        Get the status of a running test
        """
        if test_id in self.running_tests:
            return {"status": "running", "progress": 50}
        else:
            return {"status": "completed"}

# Global service instance
test_service = TestService()
