backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
PyYAML==6.0.1
aiohttp==3.8.6
